requirements.txt
pillow
pywin32
pytesseract
PyAutoGUI
PyDirectInput
fuzzywuzzy